using data_stream_ids repo instead data_points for getLatestUpdate
