+++ src/main/kotlin/dk/cachet/carp/webservices/data/repository/DataStreamIdRepository.kt
@@ -1,6 +1,8 @@
 package dk.cachet.carp.webservices.data.repository
 
 import dk.cachet.carp.webservices.data.domain.DataStreamId
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
@@ -32,7 +34,20 @@ interface DataStreamIdRepository: JpaRepository<DataStreamId, Int>
     fun deleteAllByDeploymentIds(@Param(value = "deploymentIds") deploymentIds: Collection<String>)
 
     @Query(nativeQuery = true,
-            value = "SELECT id FROM data_stream_ids WHERE study_deployment_id IN (:deploymentIds)")
+        value = "SELECT id FROM data_stream_ids WHERE study_deployment_id IN (:deploymentIds)")
     fun getAllByDeploymentIds(@Param("deploymentIds") ids: Collection<String>): List<Int>
+
+    @Query(nativeQuery = true,
+            value = "SELECT * FROM data_stream_ids WHERE study_deployment_id = :deploymentId ")
+    /*fun getAllByDeploymentIds(@Param("deploymentIds") ids: String): Collection<String>*/
+    fun getAllByDeploymentId(@Param("deploymentId") id: String, pageable: Pageable): Page<DataStreamId>
+    
+/*    fun findByDeploymentId(studyDeploymentId: String, pageable: Pageable): Page<DataStreamId>
+    @Query(nativeQuery = true,
+        value = "SELECT id FROM data_stream_ids where studyDeploymentId")*/
+
+/*    @Query(nativeQuery = true,
+        value = "SELECT up FROM data_stream_ids WHERE study_deployment_id IN (:deploymentIds)")
+    fun getAllByDeploymentIds(@Param("deploymentIds") ids: Collection<String>): List<Int>*/
 
 }
